Add web3 connectors which allow us to have an always on instance
The current file, networkConnector, is responsible for handling connections to RPC JSON endpoints. This will allow us to ensure we can always present realtime blockchain data to the end user regardless of whether or not they have a web3 wallet installed in their browser. This has been lifted from the ethersJS lib and commented to make understanding of what is happening easier for anyone who may pick this up in the future