Various hooks added to project
At present the hooks which have been created relate solely to the web3 instance which is being used within the app. In the future this will be expanded with hooks for both web3 and other things.

useEagerConnect
A hook for trying to connect to the user's injected web3 instance asap. This returns a boolean which denotes if it has already tried to connect to the user's wallet

useInactiveListener
A hook for listening to the window's ethereum instance for any events which are emitted. This is where we handle things such as the chain ID being changed, the instance disconnecting, the user changing wallet etc

useWeb3
As you would expect this is used to get the current state and lib for the web3 instance which is in use as the time of definition. The mthods available in the lib which is exposed by this component will de[end heavily on whether the user has brought an injected connector or if we are using the network connector